refactor(web): shared logic moved to handleAuthTokens
