+++ apps/web/src/features/auth/actions/signin.ts
@@ -1,21 +1,15 @@
 'use server';
 
-import { createSessionCookie } from '@/features/auth/actions/create-session-cookie';
 import { getCsrfHeaders } from '@/features/auth/actions/get-csrf-headers';
-import { post } from '@/lib/request';
-import {
-  parseCookiesFromHeader,
-  ParsedCookie,
-  setCookiesFromParsedData,
-} from '@/features/auth/utils/cookies';
+import { handleAuthTokens } from '@/features/auth/actions/handle-auth-tokens';
 import { createErrorResponse } from '@/lib/errors';
+import { post } from '@/lib/request';
 import {
   ApiResponse,
   ErrorCode,
   PublicUser,
   SigninCredentials,
 } from '@repo/types';
-import { cookies } from 'next/headers';
 
 /**
  * Server action to handle user signin
@@ -42,48 +36,11 @@ export async function signin(
       );
     }
 
-    // 4. Parse cookies from response
-    const cookieStore = await cookies();
+    // 4. Handle auth tokens and cookies
     const setCookieHeader = response.headers?.get('Set-Cookie');
-    const parsedCookies = parseCookiesFromHeader(setCookieHeader!);
+    await handleAuthTokens(setCookieHeader!, response.data);
 
-    // 5. Filter cookies and get tokens
-    let accessToken: string | undefined;
-    let refreshToken: string | undefined;
-    const filteredCookies = parsedCookies.filter((cookie: ParsedCookie) => {
-      if (cookie.name === 'refreshToken') {
-        refreshToken = cookie.value;
-        return false; // Remove from the array
-      }
-      if (cookie.name === 'accessToken') {
-        accessToken = cookie.value;
-        return true; // Keep in the array
-      }
-      return true; // Keep all other cookies
-    });
-
-    // 6. Handle no access token or refresh token
-    if (!accessToken || !refreshToken) {
-      return createErrorResponse<PublicUser>(
-        ErrorCode.UNKNOWN_ERROR,
-        'No access token or refresh token found',
-      );
-    }
-
-    // 7. Set cookies
-    setCookiesFromParsedData(cookieStore, filteredCookies);
-
-    // 8. Handle no access token
-    if (!accessToken || !refreshToken) {
-      return createErrorResponse<PublicUser>(
-        ErrorCode.UNKNOWN_ERROR,
-        'No access token or refresh token found',
-      );
-    }
-
-    // 9. Create session cookie
-    await createSessionCookie(response.data, accessToken, refreshToken);
-
+    // 5. Return the user
     return { success: true, data: response.data };
   } catch (error) {
     // Handle unexpected errors
