+++ apps/web/src/features/auth/actions/signup.ts
@@ -1,14 +1,9 @@
 'use server';
 
-import { createSessionCookie } from '@/features/auth/actions/create-session-cookie';
 import { getCsrfHeaders } from '@/features/auth/actions/get-csrf-headers';
-import { post } from '@/lib/request';
-import {
-  parseCookiesFromHeader,
-  ParsedCookie,
-  setCookiesFromParsedData,
-} from '@/features/auth/utils/cookies';
+import { handleAuthTokens } from '@/features/auth/actions/handle-auth-tokens';
 import { createErrorResponse, formatZodErrors } from '@/lib/errors';
+import { post } from '@/lib/request';
 import {
   ApiResponse,
   ErrorCode,
@@ -16,7 +11,6 @@ import {
   SignupCredentials,
 } from '@repo/types';
 import { createUserSchema } from '@repo/validation';
-import { cookies } from 'next/headers';
 
 /**
  * Server action to handle user signup
@@ -58,48 +52,11 @@ export async function signup(
       );
     }
 
-    // 6. Parse cookies from response
-    const cookieStore = await cookies();
+    // 6. Handle auth tokens and cookies
     const setCookieHeader = response.headers?.get('Set-Cookie');
-    const parsedCookies = parseCookiesFromHeader(setCookieHeader!);
+    await handleAuthTokens(setCookieHeader!, response.data);
 
-    // 7. Filter cookies and get tokens
-    let accessToken: string | undefined;
-    let refreshToken: string | undefined;
-    const filteredCookies = parsedCookies.filter((cookie: ParsedCookie) => {
-      if (cookie.name === 'refreshToken') {
-        refreshToken = cookie.value;
-        return false; // Remove from the array
-      }
-      if (cookie.name === 'accessToken') {
-        accessToken = cookie.value;
-        return true; // Keep in the array
-      }
-      return true; // Keep all other cookies
-    });
-
-    // 8. Handle no access token or refresh token
-    if (!accessToken || !refreshToken) {
-      return createErrorResponse<PublicUser>(
-        ErrorCode.UNKNOWN_ERROR,
-        'No access token or refresh token found',
-      );
-    }
-
-    // 9. Set cookies
-    setCookiesFromParsedData(cookieStore, filteredCookies);
-
-    // 10. Handle no access token
-    if (!accessToken || !refreshToken) {
-      return createErrorResponse<PublicUser>(
-        ErrorCode.UNKNOWN_ERROR,
-        'No access token or refresh token found',
-      );
-    }
-
-    // 11. Create session cookie
-    await createSessionCookie(response.data, accessToken, refreshToken);
-
+    // 7. Return the user
     return { success: true, data: response.data };
   } catch (error) {
     // Handle unexpected errors
